+++ src/core/announcements/crypto.ts
@@ -1,8 +1,57 @@
 import { ConfigOpts, requireGetSigner } from "../../config";
 import { HexString } from "../../types/Strings";
 import { ethers } from "ethers";
-import { Announcement, AnnouncementWithSignature } from "./announcementTypes";
+import {
+  Announcement,
+  BroadcastAnnouncement,
+  GraphChangeAnnouncement,
+  ProfileAnnouncement,
+  ReactionAnnouncement,
+  ReplyAnnouncement,
+} from "./factories";
 import { serialize } from "./serialization";
+
+/**
+ * A Generic Signed Announcement ready for inclusion in a Batch File
+ * {@link Announcement}
+ */
+export type AnnouncementWithSignature<T extends Announcement> = T & { signature: string };
+
+/**
+ * A Signed Announcement ready for inclusion in a Batch File
+ * {@link Announcement}
+ */
+export type SignedAnnouncement = AnnouncementWithSignature<Announcement>;
+
+/**
+ * A Signed Broadcast Announcement ready for inclusion in a Batch File
+ * {@link BroadcastAnnouncement}
+ */
+export type SignedBroadcastAnnouncement = AnnouncementWithSignature<BroadcastAnnouncement>;
+
+/**
+ * A Signed Reply Announcement ready for inclusion in a Batch File
+ * {@link ReplyAnnouncement}
+ */
+export type SignedReplyAnnouncement = AnnouncementWithSignature<ReplyAnnouncement>;
+
+/**
+ * A Signed Reaction Announcement ready for inclusion in a Batch File
+ * {@link ReactionAnnouncement}
+ */
+export type SignedReactionAnnouncement = AnnouncementWithSignature<ReactionAnnouncement>;
+
+/**
+ * A Signed Profile Announcement ready for inclusion in a Batch File
+ * {@link ProfileAnnouncement}
+ */
+export type SignedProfileAnnouncement = AnnouncementWithSignature<ProfileAnnouncement>;
+
+/**
+ * A Signed Graph Change Announcement ready for inclusion in a Batch File
+ * {@link GraphChangeAnnouncement}
+ */
+export type SignedGraphChangeAnnouncement = AnnouncementWithSignature<GraphChangeAnnouncement>;
 
 /**
  * sign() takes an Announcement and returns a Signed Announcement ready for
